+++ .template-lintrc.js
@@ -1,6 +1,3 @@
 'use strict';
 
-module.exports = {
-  plugins: ['@zestia/template-lint-config'],
-  extends: 'zestia:recommended'
-};
+module.exports = require('@zestia/template-lint-config');