+++ addon/components/select-box/input/index.js
@@ -1,7 +1,7 @@
 import Component from '@glimmer/component';
 import lifecycleActions from '../../../utils/component/lifecycle';
 import registerElement from '../../../utils/registration/element';
-import { input, keyUp } from '../../../utils/select-box/input/keyboard';
+import { input, keyDown } from '../../../utils/select-box/input/keyboard';
 import { action } from '@ember/object';
 import buildId from '../../../utils/shared/id';
 
@@ -22,7 +22,7 @@ export default class SelectBoxInput extends Component {
   }
 
   @action
-  handleKeyUp(e) {
-    keyUp(this, e);
+  handleKeyDown(e) {
+    keyDown(this, e);
   }
 }