+++ addon/utils/general/keyboard.js
@@ -3,6 +3,7 @@ export const keys = {
   9: 'tab',
   13: 'enter',
   27: 'escape',
+  32: 'space',
   37: 'left',
   38: 'up',
   39: 'right',